refactor(plugin): ♻️ drop remote provider abstract class
- SSH and FTP extend FsCrawlerExtensionFsProviderAbstract directly
- Each plugin owns parse/validate, defaults, URL scheme, and REST docs
- Keep RemoteConnectionSettings for shared job/REST/server resolution
