--- plugin/src/test/java/fr/pilato/elasticsearch/crawler/plugins/RemoteProviderCrawlerStartTest.java
@@ -1,220 +1,0 @@
-/*
- * Licensed to David Pilato (the "Author") under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. Author licenses this
- * file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- * Made from 🇫🇷🇪🇺 with ❤️ - 2011-2026
- */
-package fr.pilato.elasticsearch.crawler.plugins;
-
-import com.carrotsearch.randomizedtesting.jupiter.RandomizedTest;
-import fr.pilato.elasticsearch.crawler.fs.beans.Doc;
-import fr.pilato.elasticsearch.crawler.fs.framework.FsCrawlerIllegalConfigurationException;
-import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
-import fr.pilato.elasticsearch.crawler.fs.settings.FsSettingsLoader;
-import fr.pilato.elasticsearch.crawler.fs.test.framework.AbstractFSCrawlerTestCase;
-import java.io.InputStream;
-import java.util.LinkedHashMap;
-import java.util.Locale;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicBoolean;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-@SuppressWarnings("removal")
-class RemoteProviderCrawlerStartTest extends AbstractFSCrawlerTestCase {
-
-    @Test
-    void crawlerStartReadsFsSshWithoutValidatingAFile() {
-        String hostname = randomToken() + ".example.com";
-        String username = randomToken();
-        int port = RandomizedTest.randomIntInRange(randomizedRandomForTests, 1024, 65535);
-
-        FsSettings settings = FsSettingsLoader.load();
-        Map<String, Object> ssh = new LinkedHashMap<>();
-        ssh.put("hostname", hostname);
-        ssh.put("port", port);
-        ssh.put("username", username);
-        settings.getFs().setProviders(Map.of("ssh", ssh));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider("ssh");
-        provider.start(settings, "{}");
-
-        Assertions.assertThat(provider.getEffectiveHostname()).isEqualTo(hostname);
-        Assertions.assertThat(provider.getEffectivePort()).isEqualTo(port);
-        Assertions.assertThat(provider.getEffectiveUsername()).isEqualTo(username);
-        Assertions.assertThat(provider.validatedFile.get()).isFalse();
-        Assertions.assertThat(provider.openedConnection.get()).isFalse();
-    }
-
-    @Test
-    void crawlerStartRequiresHostname() {
-        String type = randomToken();
-        FsSettings settings = FsSettingsLoader.load();
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider(type);
-
-        Assertions.assertThatThrownBy(() -> provider.start(settings, "{}"))
-                .isInstanceOf(FsCrawlerIllegalConfigurationException.class)
-                .hasMessageContaining("fs.providers." + type + ".hostname");
-    }
-
-    @Test
-    void crawlerStartRequiresUsernameWhenNoDefault() {
-        String type = randomToken();
-        String hostname = randomToken() + ".example.com";
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getFs().setProviders(Map.of(type, Map.of("hostname", hostname)));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider(type);
-
-        Assertions.assertThatThrownBy(() -> provider.start(settings, "{}"))
-                .isInstanceOf(FsCrawlerIllegalConfigurationException.class)
-                .hasMessageContaining("fs.providers." + type + ".username");
-    }
-
-    @Test
-    void crawlerStartRejectsBlankUsername() {
-        String type = randomToken();
-        String hostname = randomToken() + ".example.com";
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getFs().setProviders(Map.of(type, Map.of("hostname", hostname, "username", "")));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider(type);
-
-        Assertions.assertThatThrownBy(() -> provider.start(settings, "{}"))
-                .isInstanceOf(FsCrawlerIllegalConfigurationException.class)
-                .hasMessageContaining("fs.providers." + type + ".username");
-    }
-
-    @Test
-    void crawlerStartDoesNotInferSshOrFtpDefaultsFromTypeName() {
-        String hostname = randomToken() + ".example.com";
-        String username = randomToken();
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getFs().setProviders(Map.of("ssh", Map.of("hostname", hostname, "username", username)));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider("ssh");
-        provider.start(settings, "{}");
-
-        Assertions.assertThat(provider.getEffectivePort()).isZero();
-        Assertions.assertThat(provider.getEffectiveUsername()).isEqualTo(username);
-        Assertions.assertThat(provider.toFileUrl("/doc.pdf")).isEqualTo("ssh://" + hostname + ":0/doc.pdf");
-    }
-
-    @Test
-    void crawlerStartDoesNotInferFtpAnonymousFromTypeName() {
-        String hostname = randomToken() + ".example.com";
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getFs().setProviders(Map.of("ftp", Map.of("hostname", hostname)));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider("ftp");
-
-        Assertions.assertThatThrownBy(() -> provider.start(settings, "{}"))
-                .isInstanceOf(FsCrawlerIllegalConfigurationException.class)
-                .hasMessageContaining("fs.providers.ftp.username");
-    }
-
-    @Test
-    void crawlerStartUsesProviderDefaultPortAndUsername() {
-        String hostname = randomToken() + ".example.com";
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getFs().setProviders(Map.of("ftp", Map.of("hostname", hostname)));
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider("ftp", 21, "anonymous");
-        provider.start(settings, "{}");
-
-        Assertions.assertThat(provider.getEffectiveUsername()).isEqualTo("anonymous");
-        Assertions.assertThat(provider.getEffectivePort()).isEqualTo(21);
-    }
-
-    @Test
-    void crawlerStartFallsBackToDeprecatedServer() {
-        String hostname = randomToken() + ".example.com";
-        FsSettings settings = FsSettingsLoader.load();
-        settings.getServer().setHostname(hostname);
-        settings.getServer().setUsername(randomToken());
-
-        RecordingRemoteProvider provider = new RecordingRemoteProvider("ssh");
-        provider.start(settings, "{}");
-
-        Assertions.assertThat(provider.getEffectiveHostname()).isEqualTo(hostname);
-        Assertions.assertThat(provider.validatedFile.get()).isFalse();
-    }
-
-    private String randomToken() {
-        return RandomizedTest.randomAsciiLettersOfLengthBetween(randomizedRandomForTests, 6, 12)
-                .toLowerCase(Locale.ROOT);
-    }
-
-    private static final class RecordingRemoteProvider extends FsCrawlerExtensionRemoteProviderAbstract {
-        private final String type;
-        private final int defaultPort;
-        private final String defaultUsername;
-        private final AtomicBoolean validatedFile = new AtomicBoolean();
-        private final AtomicBoolean openedConnection = new AtomicBoolean();
-
-        private RecordingRemoteProvider(String type) {
-            this(type, 0, null);
-        }
-
-        private RecordingRemoteProvider(String type, int defaultPort, String defaultUsername) {
-            this.type = type;
-            this.defaultPort = defaultPort;
-            this.defaultUsername = defaultUsername;
-        }
-
-        @Override
-        public String getType() {
-            return type;
-        }
-
-        @Override
-        protected int defaultPort() {
-            return defaultPort;
-        }
-
-        @Override
-        protected String defaultUsername() {
-            return defaultUsername;
-        }
-
-        @Override
-        protected long getFilesize() {
-            return 0;
-        }
-
-        @Override
-        protected void doValidateFile() {
-            validatedFile.set(true);
-        }
-
-        @Override
-        public void openConnection() {
-            openedConnection.set(true);
-        }
-
-        @Override
-        public InputStream readFile() {
-            return InputStream.nullInputStream();
-        }
-
-        @Override
-        public Doc createDocument() {
-            return new Doc();
-        }
-    }
-}
