+++ plugins/fs-ftp-plugin/src/main/java/fr/pilato/elasticsearch/crawler/plugins/fs/ftp/FsFtpPlugin.java
@@ -20,11 +20,13 @@
  */
 package fr.pilato.elasticsearch.crawler.plugins.fs.ftp;
 
+import fr.pilato.elasticsearch.crawler.fs.beans.Doc;
 import fr.pilato.elasticsearch.crawler.fs.beans.FileAbstractModel;
 import fr.pilato.elasticsearch.crawler.fs.framework.FsCrawlerUtil;
-import fr.pilato.elasticsearch.crawler.plugins.FsCrawlerExtensionRemoteProviderAbstract;
+import fr.pilato.elasticsearch.crawler.plugins.FsCrawlerExtensionFsProviderAbstract;
 import fr.pilato.elasticsearch.crawler.plugins.FsCrawlerPlugin;
 import fr.pilato.elasticsearch.crawler.plugins.FsCrawlerPluginException;
+import fr.pilato.elasticsearch.crawler.plugins.RemoteConnectionSettings;
 import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -60,7 +62,7 @@ public class FsFtpPlugin extends FsCrawlerPlugin {
     }
 
     @Extension
-    public static class FsCrawlerExtensionFsProviderFtp extends FsCrawlerExtensionRemoteProviderAbstract {
+    public static class FsCrawlerExtensionFsProviderFtp extends FsCrawlerExtensionFsProviderAbstract {
         /** Default FTP port when {@code fs.providers.ftp.port} is omitted. */
         public static final int DEFAULT_PORT = 21;
 
@@ -77,9 +79,13 @@ public class FsFtpPlugin extends FsCrawlerPlugin {
 
         private FTPClient ftp;
         private boolean isUtf8 = false;
-
-        // FTP-specific fields
         private FTPFile fileInfo;
+
+        private String remotePath;
+        private String hostname;
+        private int port;
+        private String username;
+        private String password;
 
         private final Predicate<FTPFile> isNotSymLink = file -> {
             if (fsSettings != null && fsSettings.getFs().isFollowSymlinks()) return true;
@@ -92,24 +98,108 @@ public class FsFtpPlugin extends FsCrawlerPlugin {
         }
 
         @Override
-        protected int defaultPort() {
-            return DEFAULT_PORT;
+        public boolean supportsCrawling() {
+            return true;
         }
 
         @Override
-        protected String defaultUsername() {
-            return DEFAULT_USERNAME;
+        protected void parseSettings() {
+            RemoteConnectionSettings resolved =
+                    RemoteConnectionSettings.resolve(getType(), document, fsSettings, DEFAULT_PORT, DEFAULT_USERNAME);
+            remotePath = resolved.remotePath();
+            hostname = resolved.hostname();
+            port = resolved.port();
+            username = resolved.username();
+            password = resolved.password();
+            resolved.deprecationWarnings().forEach(logger::warn);
         }
 
-        // ========== Protocol-specific settings ==========
-
         @Override
-        protected long getFilesize() {
-            return fileInfo != null ? fileInfo.getSize() : 0;
+        protected void validateSettings() throws IOException {
+            if (hostname == null || hostname.isBlank()) {
+                throw new IOException("Provider [" + getType() + "] requires fs.providers." + getType() + ".hostname");
+            }
+            if (username == null || username.isBlank()) {
+                throw new IOException("Provider [" + getType() + "] requires fs.providers." + getType() + ".username");
+            }
+
+            if (remotePath == null || remotePath.isEmpty()) {
+                if (document != null) {
+                    throw new IOException(getType() + " path is missing");
+                }
+                return;
+            }
+
+            remotePath = normalizeRemotePath(remotePath);
+
+            boolean success = false;
+            try {
+                openConnection();
+                validateRemoteFile();
+                success = true;
+            } catch (FsCrawlerPluginException e) {
+                throw e;
+            } catch (Exception e) {
+                throw new FsCrawlerPluginException(
+                        "Failed to connect to " + getType().toUpperCase() + " server: " + e.getMessage(), e);
+            } finally {
+                if (!success) {
+                    try {
+                        closeConnection();
+                    } catch (Exception e) {
+                        logger.warn(
+                                "Error closing {} connection after validation failure: {}",
+                                getType().toUpperCase(),
+                                e.getMessage());
+                    }
+                }
+            }
         }
 
         @Override
-        protected void doValidateFile() throws FsCrawlerPluginException {
+        public Doc createDocument() {
+            String filename = FilenameUtils.getName(remotePath);
+            logger.debug("Creating document from {} for file {}", getType(), filename);
+
+            Doc doc = new Doc();
+            doc.getFile().setFilename(filename);
+            doc.getFile().setFilesize(fileInfo != null ? fileInfo.getSize() : 0);
+
+            String rootUrl = (fsSettings.getFs() != null && fsSettings.getFs().getUrl() != null)
+                    ? fsSettings.getFs().getUrl()
+                    : "/";
+            doc.getPath().setVirtual(FsCrawlerUtil.computeVirtualPathName(rootUrl, remotePath));
+            doc.getPath().setReal(remotePath);
+            return doc;
+        }
+
+        @Override
+        public void stop() throws FsCrawlerPluginException {
+            closeConnection();
+        }
+
+        @Override
+        public String toFileUrl(String fullPath) {
+            return String.format("ftp://%s:%d%s", hostname, port, fullPath);
+        }
+
+        private String normalizeRemotePath(String path) throws IOException {
+            if (path == null) {
+                return null;
+            }
+            if (!path.startsWith("/")) {
+                String rootPath =
+                        fsSettings.getFs() != null ? fsSettings.getFs().getUrl() : null;
+                if (rootPath == null || rootPath.isEmpty()) {
+                    throw new IOException("Cannot resolve relative path [" + path + "]: fs.url is not configured. "
+                            + "Please use an absolute path starting with '/' or configure fs.url in the job settings.");
+                }
+                return rootPath.endsWith("/") ? rootPath + path : rootPath + "/" + path;
+            }
+            return path;
+        }
+
+        private void validateRemoteFile() throws FsCrawlerPluginException {
             try {
                 // Get file info to validate it exists
                 // Use mlistFile() which returns info about the path itself (like SSH stat())
@@ -247,25 +337,19 @@ public class FsFtpPlugin extends FsCrawlerPlugin {
                 // Send a safe command (NOOP) over the control connection to reset the router's idle timer
                 ftp.setControlKeepAliveTimeout(Duration.ofSeconds(300));
 
-                String effectiveHostname = getEffectiveHostname();
-                int effectivePort = getEffectivePort();
-                String effectiveUsername = getEffectiveUsername();
-                String effectivePassword = getEffectivePassword();
+                logger.debug("Opening FTP connection to {}@{}", username, hostname);
 
-                logger.debug("Opening FTP connection to {}@{}", effectiveUsername, effectiveHostname);
-
-                ftp.connect(effectiveHostname, effectivePort);
+                ftp.connect(hostname, port);
 
                 // Check FTP client connection
                 int reply = ftp.getReplyCode();
                 if (!FTPReply.isPositiveCompletion(reply)) {
                     ftp.disconnect();
-                    logger.warn("Cannot connect with FTP to {}@{}", effectiveUsername, effectiveHostname);
-                    throw new FsCrawlerPluginException(
-                            "Can not connect to " + effectiveUsername + "@" + effectiveHostname);
+                    logger.warn("Cannot connect with FTP to {}@{}", username, hostname);
+                    throw new FsCrawlerPluginException("Can not connect to " + username + "@" + hostname);
                 }
 
-                if (!ftp.login(effectiveUsername, effectivePassword)) {
+                if (!ftp.login(username, password)) {
                     ftp.disconnect();
                     throw new FsCrawlerPluginException("Please check ftp user or password");
                 }
